made the neighbors method its own class in map utilities
