+++ src/Utilities/AIUtilities/Astar.java
@@ -2,6 +2,7 @@ package Utilities.AIUtilities;
 
 import Model.Location;
 import Model.Map.Tile;
+import Utilities.MapUtilities.Neighbors;
 import Utilities.Settings;
 import Model.Map.Map;
 
@@ -12,7 +13,7 @@ import java.util.*;
  */
 
 public class Astar {
-
+    
     public static ArrayList<Location> Findpath(Map map, Location start, Location end) {
 
         /*
@@ -29,7 +30,7 @@ public class Astar {
             Location current = Queue.remove();
             Tile currentTile = map.getTile(current.getX(),current.getY());
             if(currentTile.isWalkable() || start.equals(current)) {
-                Tile neighbors[] = neighbors(currentTile, map);   //This uses the neighbor method which calculates all the
+                Tile neighbors[] = Neighbors.neighbors(currentTile, map);   //This uses the neighbor method which calculates all the
                 for (int i = 0; i < 6; i++) {                     //surrounding tiles of a particular tile.
                     Tile next = neighbors[i];
                     if (next != null && parent[next.getLocation().getX()][next.getLocation().getY()] == null && !next.getLocation().equals(start)) { //This if statement makes sure that the tile isnt null it is not already a parent and it's not the starting tile
@@ -103,26 +104,4 @@ public class Astar {
         return path;
     }
 
-    public static Tile[] neighbors(Tile start, Map map) {
-        int x = start.getLocation().getX();
-        int y = start.getLocation().getY();
-        Tile neighbors[] = new Tile[6];
-        if (x % 2 == 0) {
-            neighbors[0] = map.getTile(x,y - 1);
-            neighbors[1] = map.getTile(x + 1,y - 1);
-            neighbors[2] = map.getTile(x + 1,y);
-            neighbors[3] = map.getTile(x,y + 1);
-            neighbors[4] = map.getTile(x - 1,y);
-            neighbors[5] = map.getTile(x - 1,y - 1);
-        }
-        else{
-            neighbors[0] = map.getTile(x,y - 1);
-            neighbors[1] = map.getTile(x + 1,y);
-            neighbors[2] = map.getTile(x + 1,y + 1);
-            neighbors[3] = map.getTile(x ,y + 1);
-            neighbors[4] = map.getTile(x - 1,y + 1);
-            neighbors[5] = map.getTile(x - 1,y);
-        }
-        return neighbors;
-    }
 }
